# Conflicts:
#	build_scripts/build_usd.py
#	cmake/defaults/Packages.cmake
#	pxr/imaging/garch/baseTextureData.h
#	pxr/imaging/garch/ptexMipmapTextureLoader.cpp
#	pxr/imaging/garch/uvTexture.cpp
#	pxr/imaging/garch/uvTextureData.cpp
#	pxr/imaging/garch/vdbTextureData.cpp
#	pxr/imaging/glf/CMakeLists.txt
#	pxr/imaging/glf/baseTexture.cpp
#	pxr/imaging/glf/contextCaps.cpp
#	pxr/imaging/glf/glew.cpp
#	pxr/imaging/glf/module.cpp
#	pxr/imaging/glf/plugInfo.json
#	pxr/imaging/glf/ptexMipmapTextureLoader.cpp
#	pxr/imaging/glf/ptexTexture.cpp
#	pxr/imaging/glf/ptexTexture.h
#	pxr/imaging/glf/udimTexture.cpp
#	pxr/imaging/glf/udimTexture.h
#	pxr/imaging/hd/textureResource.h
#	pxr/imaging/hdSt/CMakeLists.txt
#	pxr/imaging/hdSt/codeGen.cpp
#	pxr/imaging/hdSt/extCompGpuComputationResource.cpp
#	pxr/imaging/hdSt/geometricShader.cpp
#	pxr/imaging/hdSt/glfTextureCpuData.cpp
#	pxr/imaging/hdSt/immediateDrawBatch.cpp
#	pxr/imaging/hdSt/material.cpp
#	pxr/imaging/hdSt/materialBufferSourceAndTextureHelper.cpp
#	pxr/imaging/hdSt/materialNetwork.cpp
#	pxr/imaging/hdSt/ptexMipmapTextureLoader.cpp
#	pxr/imaging/hdSt/ptexMipmapTextureLoader.h
#	pxr/imaging/hdSt/ptexTextureObject.cpp
#	pxr/imaging/hdSt/ptexTextureObject.h
#	pxr/imaging/hdSt/renderPass.cpp
#	pxr/imaging/hdSt/renderPassState.cpp
#	pxr/imaging/hdSt/resourceBinder.cpp
#	pxr/imaging/hdSt/resourceRegistry.cpp
#	pxr/imaging/hdSt/samplerObject.cpp
#	pxr/imaging/hdSt/samplerObject.h
#	pxr/imaging/hdSt/shaderCode.h
#	pxr/imaging/hdSt/shaders/domeLight.glslfx
#	pxr/imaging/hdSt/shaders/meshNormal.glslfx
#	pxr/imaging/hdSt/shaders/ptexTexture.glslfx
#	pxr/imaging/hdSt/surfaceShader.cpp
#	pxr/imaging/hdSt/texture.cpp
#	pxr/imaging/hdSt/textureBinder.cpp
#	pxr/imaging/hdSt/textureObject.cpp
#	pxr/imaging/hdSt/textureObject.h
#	pxr/imaging/hdSt/textureResource.cpp
#	pxr/imaging/hdSt/textureResource.h
#	pxr/imaging/hdSt/udimTextureObject.cpp
#	pxr/imaging/hdSt/udimTextureObject.h
#	pxr/imaging/hdx/shadowTask.cpp
#	pxr/imaging/hdx/taskController.h
#	pxr/imaging/hgiMetal/shaderFunction.mm
#	pxr/usd/usdLux/discoveryPlugin.h
#	pxr/usdImaging/usdImagingGL/engine.cpp
#	pxr/usdImaging/usdImagingGL/unitTestGLDrawing.cpp

+++ pxr/imaging/hdSt/ptexMipmapTextureLoader.cpp
@@ -48,7 +48,7 @@ HdStPtexMipmapTextureLoader::Block::guttering(
     int lineBufferSize = std::max(wid, hei) * bpp;
     unsigned char * lineBuffer = new unsigned char[lineBufferSize];
 
-    int numEdges = ptex->meshType() == Ptex::mt_triangle ? 3 : 4;  
+    int numEdges = ptex->meshType() == Ptex::mt_triangle ? 3 : 4;
 
     for (int edge = 0; edge < numEdges; edge++) {
         int len = (edge == 0 || edge == 2) ? wid : hei;
@@ -231,7 +231,7 @@ struct HdStPtexMipmapTextureLoader::Page
     {
         explicit SlotLimit(size_t num, uint16_t w, uint16_t h) :
             numTexels(num), width(w), height(h) { }
-        
+
         const uint32_t numTexels;
         const uint16_t width, height;
     };
@@ -406,7 +406,7 @@ public:
                 _currentFace = _ptex->getFaceInfo(_currentFace).adjface(2);
                 _currentEdge = 1;
                 _mid = false;
-            } else if (info.isSubface() && 
+            } else if (info.isSubface() &&
                 (!_ptex->getFaceInfo(_currentFace).isSubface()) &&
                 _currentEdge == 3) {
                 _mid = true;
@@ -436,7 +436,7 @@ public:
             }
         }
         Ptex::FaceInfo nextFaceInfo = _ptex->getFaceInfo(_currentFace);
-        if ((!_clockWise) && 
+        if ((!_clockWise) &&
             (!info.isSubface()) && (nextFaceInfo.isSubface())) {
              // needs tricky traverse for boundary subface...
              if (_currentEdge == 3) {
@@ -578,7 +578,7 @@ HdStPtexMipmapTextureLoader::resampleBorder(int face, int edgeId,
             --res.vlog2;
             srcLength /= 2;
         }
-        
+
         PtexFaceData * data = _ptex->getData(face, res);
         if (!data) {
             // XXX:validation
@@ -934,7 +934,7 @@ HdStPtexMipmapTextureLoader::optimizePacking(int maxNumPages,
         smallestBlockWidth = std::min(smallestBlockWidth, (*it)->width);
         smallestBlockHeight = std::min(smallestBlockHeight, (*it)->height);
     }
-    
+
     // compute page size ---------------------------------------------
     {
         // page size is set to the largest edge of the largest block :
@@ -957,7 +957,7 @@ HdStPtexMipmapTextureLoader::optimizePacking(int maxNumPages,
         if (w < minPageSize) {
             w = minPageSize;
         }
-        
+ 
         if (h < minPageSize) {
             h = minPageSize;
         }
@@ -1068,7 +1068,7 @@ HdStPtexMipmapTextureLoader::generateBuffers()
     int numPages = (int)_pages.size();
 
     // populate the texels
-    size_t pageStride = _bpp * _pageWidth * _pageHeight;
+    int pageStride = _bpp * _pageWidth * _pageHeight;
 
     _texelBuffer = new unsigned char[pageStride * numPages];
     _memoryUsage = pageStride * numPages;
